requirements.txt
xlrd==1.2.0
requests~=2.25.1
tldextract~=3.1.0
numpy~=1.23.5
Pillow~=8.0.1
pandas~=1.5.2
XlsxWriter~=1.4.3
tools~=0.1.9
ipywidgets~=7.5.1
ipython~=7.18.1
cryptography~=37.0.2
glob2~=0.7
humanize~=4.2.1
psutil~=5.9.1
future~=0.18.2
coloredlogs~=15.0
colorama~=0.4.4
termcolor~=1.1.0
pytz~=2021.1
tzlocal~=2.1
matplotlib~=3.3.2
scipy==1.9.3
Deprecated~=1.2.10
toolz~=0.11.1
lxml~=4.6.3
selection~=0.0.14
weblib~=0.1.30
pycryptodome==3.9.9

pypiwin32==223
pywin32==301